+++ filament/backend/src/vulkan/VulkanFboCache.cpp
@@ -120,6 +120,7 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
         return iter->second.handle;
     }
     const bool isSwapChain = config.colorLayout[0] == VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
+    const bool hasSubpasses = config.subpassMask != 0;
 
     // Set up some const aliases for terseness.
     const VkAttachmentLoadOp kClear = VK_ATTACHMENT_LOAD_OP_CLEAR;
@@ -138,7 +139,12 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
     struct { VkImageLayout subpass, initial, final; } colorLayouts[MRT::TARGET_COUNT];
     if (isSwapChain) {
         colorLayouts[0].subpass = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-        colorLayouts[0].initial = discard ? VK_IMAGE_LAYOUT_UNDEFINED : colorLayouts[0].subpass;
+
+        // It is legal to always use UNDEFINED for "initial", but we wish to avoid warnings
+        // when the load op is LOAD.
+        colorLayouts[0].initial = discard ? VK_IMAGE_LAYOUT_UNDEFINED :
+                VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
+
         colorLayouts[0].final = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
     } else {
         for (int i = 0; i < MRT::TARGET_COUNT; i++) {
@@ -149,7 +155,7 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
     }
 
     VkAttachmentReference inputAttachmentRef[MRT::TARGET_COUNT] = {};
-    VkAttachmentReference colorAttachmentRef[MRT::TARGET_COUNT] = {};
+    VkAttachmentReference colorAttachmentRefs[2][MRT::TARGET_COUNT] = {};
     VkAttachmentReference resolveAttachmentRef[MRT::TARGET_COUNT] = {};
     VkAttachmentReference depthAttachmentRef = {};
 
@@ -157,14 +163,15 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
 
     VkSubpassDescription subpasses[2] = {{
         .pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS,
-        .pColorAttachments = colorAttachmentRef,
+        .pInputAttachments = nullptr,
+        .pColorAttachments = colorAttachmentRefs[0],
         .pResolveAttachments = resolveAttachmentRef,
         .pDepthStencilAttachment = hasDepth ? &depthAttachmentRef : nullptr
     },
     {
         .pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS,
         .pInputAttachments = inputAttachmentRef,
-        .pColorAttachments = colorAttachmentRef,
+        .pColorAttachments = colorAttachmentRefs[1],
         .pResolveAttachments = resolveAttachmentRef,
         .pDepthStencilAttachment = hasDepth ? &depthAttachmentRef : nullptr
     }};
@@ -173,24 +180,6 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
     // For simplicity, create an array that can hold the maximum possible number of attachments.
     // Note that this needs to have the same ordering as the corollary array in getFramebuffer.
     VkAttachmentDescription attachments[MRT::TARGET_COUNT + MRT::TARGET_COUNT + 1] = {};
-
-    // Determine the number of color attachments based on whether the format has been initialized.
-    int colorAttachmentCount = 0;
-    for (VkFormat format : config.colorFormat) {
-        if (format != VK_FORMAT_UNDEFINED) {
-            ++colorAttachmentCount;
-        }
-    }
-    subpasses[0].colorAttachmentCount = colorAttachmentCount;
-    subpasses[1].colorAttachmentCount = colorAttachmentCount;
-
-    // Nulling out the zero-sized lists is necessary to avoid VK_ERROR_OUT_OF_HOST_MEMORY on Adreno.
-    if (colorAttachmentCount == 0) {
-        subpasses[0].pColorAttachments = nullptr;
-        subpasses[0].pResolveAttachments = nullptr;
-        subpasses[1].pColorAttachments = nullptr;
-        subpasses[1].pResolveAttachments = nullptr;
-    }
 
     // We support 2 subpasses, which means we need to supply 1 dependency struct.
     VkSubpassDependency dependencies[1] = {{
@@ -207,32 +196,56 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
         .sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO,
         .attachmentCount = 0u,
         .pAttachments = attachments,
-        .subpassCount = config.subpassMask ? 2u : 1u,
+        .subpassCount = hasSubpasses ? 2u : 1u,
         .pSubpasses = subpasses,
-        .dependencyCount = config.subpassMask ? 1u : 0u,
+        .dependencyCount = hasSubpasses ? 1u : 0u,
         .pDependencies = dependencies
     };
 
     int attachmentIndex = 0;
 
     // Populate the Color Attachments.
-    VkAttachmentReference* pColorAttachment = colorAttachmentRef;
     for (int i = 0; i < MRT::TARGET_COUNT; i++) {
         if (config.colorFormat[i] == VK_FORMAT_UNDEFINED) {
             continue;
         }
-        TargetBufferFlags flag = TargetBufferFlags(int(TargetBufferFlags::COLOR0) << i);
-        bool clear = any(config.clear & flag);
-        bool discard = any(config.discardStart & flag);
-        if (config.subpassMask & (1 << i)) {
-            int subpassInputIndex = subpasses[1].inputAttachmentCount++;
-            inputAttachmentRef[subpassInputIndex].layout = colorLayouts[i].subpass;
-            inputAttachmentRef[subpassInputIndex].attachment = attachmentIndex;
+        const VkImageLayout subpassLayout = colorLayouts[i].subpass;
+        uint32_t index;
+
+        if (!hasSubpasses) {
+            index = subpasses[0].colorAttachmentCount++;
+            colorAttachmentRefs[0][index].layout = subpassLayout;
+            colorAttachmentRefs[0][index].attachment = attachmentIndex;
+        } else {
+
+            // The Driver API consolidates all color attachments from the first and second subpasses
+            // into a single list, and uses a bitmask to mark attachments that belong only to the
+            // second subpass and should be available as inputs. All color attachments in the first
+            // subpass are automatically made available to the second subpass.
+
+            // If there are subpasses, we require the input attachment to be the first attachment.
+            // Breaking this assumption would likely require enhancements to the Driver API in order
+            // to supply Vulkan with all the information needed.
+            assert(config.subpassMask == 1);
+
+            if (config.subpassMask & (1 << i)) {
+                index = subpasses[0].colorAttachmentCount++;
+                colorAttachmentRefs[0][index].layout = subpassLayout;
+                colorAttachmentRefs[0][index].attachment = attachmentIndex;
+
+                index = subpasses[1].inputAttachmentCount++;
+                inputAttachmentRef[index].layout = subpassLayout;
+                inputAttachmentRef[index].attachment = attachmentIndex;
+            }
+
+            index = subpasses[1].colorAttachmentCount++;
+            colorAttachmentRefs[1][index].layout = subpassLayout;
+            colorAttachmentRefs[1][index].attachment = attachmentIndex;
         }
 
-        pColorAttachment->layout = colorLayouts[i].subpass;
-        pColorAttachment->attachment = attachmentIndex;
-        ++pColorAttachment;
+        const TargetBufferFlags flag = TargetBufferFlags(int(TargetBufferFlags::COLOR0) << i);
+        const bool clear = any(config.clear & flag);
+        const bool discard = any(config.discardStart & flag);
 
         attachments[attachmentIndex++] = {
             .format = config.colorFormat[i],
@@ -244,6 +257,14 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
             .initialLayout = colorLayouts[i].initial,
             .finalLayout = colorLayouts[i].final
         };
+    }
+
+    // Nulling out the zero-sized lists is necessary to avoid VK_ERROR_OUT_OF_HOST_MEMORY on Adreno.
+    if (subpasses[0].colorAttachmentCount == 0) {
+        subpasses[0].pColorAttachments = nullptr;
+        subpasses[0].pResolveAttachments = nullptr;
+        subpasses[1].pColorAttachments = nullptr;
+        subpasses[1].pResolveAttachments = nullptr;
     }
 
     // Populate the Resolve Attachments.
@@ -304,7 +325,7 @@ VkRenderPass VulkanFboCache::getRenderPass(RenderPassKey config) noexcept {
     utils::slog.d << "Created render pass " << renderPass << " with "
         << "samples = " << int(config.samples) << ", "
         << "depth = " << (hasDepth ? 1 : 0) << ", "
-        << "colorAttachmentCount = " << colorAttachmentCount
+        << "colorAttachmentCount[0] = " << subpasses[0].colorAttachmentCount
         << utils::io::endl;
     #endif