+++ libs/image/include/image/ColorTransform.h
@@ -347,7 +347,7 @@ inline LinearImage fromLinearToRGBM(const LinearImage& image) {
 }
 
 template<typename T>
-static LinearImage toLinear(size_t w, size_t h, size_t bpr, const uint8_t* src) {
+static LinearImage toLinearWithAlpha(size_t w, size_t h, size_t bpr, const uint8_t* src) {
     LinearImage result(w, h, 4);
     filament::math::float4* d = reinterpret_cast<filament::math::float4*>(result.getPixelRef(0, 0));
     for (size_t y = 0; y < h; ++y) {
@@ -356,6 +356,21 @@ static LinearImage toLinear(size_t w, size_t h, size_t bpr, const uint8_t* src) 
             filament::math::float3 sRGB(p[0], p[1], p[2]);
             sRGB /= std::numeric_limits<T>::max();
             *d++ = filament::math::float4(sRGBToLinear(sRGB), 1.0f);
+        }
+    }
+    return result;
+}
+
+template<typename T>
+static LinearImage toLinear(size_t w, size_t h, size_t bpr, const uint8_t* src) {
+    LinearImage result(w, h, 3);
+    filament::math::float3* d = reinterpret_cast<filament::math::float3*>(result.getPixelRef(0, 0));
+    for (size_t y = 0; y < h; ++y) {
+        T const* p = reinterpret_cast<T const*>(src + y * bpr);
+        for (size_t x = 0; x < w; ++x, p += 3) {
+            filament::math::float3 sRGB(p[0], p[1], p[2]);
+            sRGB /= std::numeric_limits<T>::max();
+            *d++ = sRGBToLinear(sRGB);
         }
     }
     return result;